requirements.txt
django==4.2.16
djangorestframework==3.15.2
django-cors-headers==4.5.0
djangorestframework-simplejwt==5.3.1
python-dotenv==1.0.1
pillow==10.4.0          
qrcode[pil]==8.0      